Renamed so3healpix to so3data to add array later
